スクロールの制御とレイアウトの修正
