+++ src/components/Header/Header.css
@@ -3,7 +3,7 @@ header {
   top: 0;
   left: 0;
   width: 100%;
-  background: rgba(82, 179, 217, 0.9);
+  background: rgba(82, 179, 217);
   height: 56px;
 }
