+++ src/containers/ChatList/ChatList.css
@@ -1,0 +1,10 @@
+.chat-list-wrapper {
+  margin-top: 65px;
+  margin-bottom: 65px;
+  width: 100%;
+  height: 100%;
+}
+
+.message-end {
+  height: 5px;
+}
