--- src/containers/Main/Main.css
@@ -1,5 +1,0 @@
-.chat-list-wrapper {
-  margin-top: 65px;
-  width: 100%;
-  height: 100%;
-}
